工作区整备：清除sh_2.0/sh_3.0旧版整仓与根目录历史文档，六仓（joint/face/face-LRU/wscllm/wscllm-LRU）代码同步更新，joint仓新增KV配额/远端信用/SLO域指标等测试与工具，论文素材补充（D2D调研、远端共享KV池设计、novelty检索等）（2026-09-24）

+++ astra-sim-face-LRU/astra-sim/system/SimSendCaller.cc
@@ -10,7 +10,7 @@ using namespace AstraSim;
 
 SimSendCaller::SimSendCaller(Sys* sys,
                              void* buffer,
-                             int count,
+                             uint64_t count,
                              int type,
                              int dst,
                              int tag,
